finishing test

+++ src/main/kotlin/model/entity/Transaction.kt
@@ -1,13 +1,12 @@
 package model.entity
 
-import model.resources.Currency
 import java.math.BigDecimal
 import java.time.Instant
 
 data class Transaction(
     val sourceCurrency: String,
-    val sourceValue: BigDecimal,
+    val sourceValue: Double,
     val destinationCurrency: String,
-    val conversionRate: BigDecimal,
+    val conversionRate: Double,
     val date: Instant = Instant.now()
 )
